+++ CHANGES.md
@@ -66,6 +66,7 @@
 
 - IF: Complete code overhaul, with some syntax changes
 - IF: This program is tolerated as completely independent program from the shell
+- HELPFUL: split_commands() separates commands not only after keywords requested in args, but also when '\n' char is found
 
 ### Waiting features:
 
@@ -73,7 +74,6 @@
 - WHILE, LOOP, MATCH, FOR: More logic operations
 - BREAK: Quit from the loop
 - IF: Add support for ELSE and ELSEIF (again)
-- HELPFUL: split_commands() separates commands not only after keywords requested in args, but also when '\n' char is found
 
 # Release 0.4.0: